+++ relay-spans/src/lib.rs
@@ -11,6 +11,7 @@ pub use crate::name::name_for_attributes;
 pub use crate::op::derive_op_for_v2_span;
 pub use crate::otel_to_sentry_v2::otel_to_sentry_span as otel_to_sentry_span_v2;
 pub use crate::v1_to_v2::span_v1_to_span_v2;
+pub use crate::web_vitals::extract_web_vital_metrics;
 
 pub use opentelemetry_proto::tonic::trace::v1 as otel_trace;
 
@@ -19,3 +20,4 @@ mod name;
 mod op;
 mod otel_to_sentry_v2;
 mod v1_to_v2;
+mod web_vitals;