+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 **Features**:
 
+- Emit web-vitals as metrics. ([#6118](https://github.com/getsentry/relay/pull/6118))
 - No longer write the deprecated `sentry.transaction` and `db.system` attributes. ([#6237](https://github.com/getsentry/relay/pull/6237), [#6238](https://github.com/getsentry/relay/pull/6238))
 - Allow additional exceptions in minidump and apple crash report events. ([#6241](https://github.com/getsentry/relay/pull/6241))